Add `Game` abstraction with `perform_move`
Adds the new `Game` struct which represents a ongoing game of chess.
This meant moving the information of the current player out of the
`Board` struct into `Game`.

I also had to change how checkmate is represented. Honestly can't
remember exactly what was changed, and can't be bothered to check. :)

Also, instances of `CheckMate` was corrected to `Checkmate`, as
checkmate is spelled as one word.

Resolves #1