Add `moves` library
Adds the new private `moves` library, which contains a collection of
helper functions for generating lists of valid moves. They give the list
of moves for a piece at a position assuming that the board is otherwise
empty.

These will be ran through more integrated functions which validate these
moves.