Fix bug where king could block check mate
Fixes a bug where if a king in check could move so that it's old
position was in between the attackers and a new empty position it
that attacker would be considered blocked.

See the new test for an example where this was the case.

This required a new BoardView abstraction.

Also, unfortunately some unrelated formatting changes snuck in to this
commit. I don't have time to rebase them in properly.